container/requirements.txt
fastapi
uvicorn
torch==2.4.1
torchvision==0.19.1
pillow
python-multipart